# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Modular API server. Run in platform or tenant mode via -mode flag.",
        "title": "Go Web API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/billing/payments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns paginated payment history for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "List payment history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_billing_models.PaymentListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Records a payment against an active subscription.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Record a payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Payment details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_billing_models.RecordPaymentDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_billing_models.Payment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/billing/plan": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the authenticated user's active plan, subscription status, expiry date, and all feature limits.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Get current plan info and feature limits",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_billing_models.PlanInfo"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/billing/subscription": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the authenticated user's active subscription, or null if none.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Get current subscription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_billing_models.Subscription"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new subscription or upgrades the existing one. Cancels the old subscription when switching plans.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Billing"
                ],
                "summary": "Create or upgrade a subscription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Subscription details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_billing_models.UpsertSubscriptionDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_billing_models.Subscription"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/josex_web_modules_core_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "josex_web_modules_billing_models.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "paid_at": {
                    "type": "string"
                },
                "period_end": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "provider_payment_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subscription_id": {
                    "type": "string"
                }
            }
        },
        "josex_web_modules_billing_models.PaymentListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/josex_web_modules_billing_models.Payment"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "josex_web_modules_billing_models.PlanInfo": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_expired": {
                    "type": "boolean"
                },
                "limits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/josex_web_modules_billing_models.PlanLimit"
                    }
                },
                "plan": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "josex_web_modules_billing_models.PlanLimit": {
            "type": "object",
            "properties": {
                "feature": {
                    "type": "string"
                },
                "limit_value": {
                    "description": "-1 = unlimited",
                    "type": "integer"
                }
            }
        },
        "josex_web_modules_billing_models.RecordPaymentDto": {
            "type": "object",
            "required": [
                "amount",
                "subscription_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "period_end": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "provider_payment_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "completed",
                        "failed",
                        "refunded"
                    ]
                },
                "subscription_id": {
                    "type": "string"
                }
            }
        },
        "josex_web_modules_billing_models.Subscription": {
            "type": "object",
            "properties": {
                "canceled_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "plan": {
                    "description": "free | pro | enterprise",
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "description": "active | expired | canceled | trial",
                    "type": "string"
                },
                "stripe_customer_id": {
                    "type": "string"
                },
                "stripe_sub_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "josex_web_modules_billing_models.UpsertSubscriptionDto": {
            "type": "object",
            "required": [
                "plan"
            ],
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "plan": {
                    "type": "string",
                    "enum": [
                        "free",
                        "pro",
                        "enterprise"
                    ]
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "expired",
                        "canceled",
                        "trial"
                    ]
                },
                "stripe_customer_id": {
                    "type": "string"
                },
                "stripe_sub_id": {
                    "type": "string"
                }
            }
        },
        "josex_web_modules_core_errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code\nexample: 400",
                    "type": "string"
                },
                "detail": {
                    "description": "Error details\nexample: {\"email\": \"Invalid email format\"}"
                },
                "error": {
                    "description": "Error message\nexample: Bad Request"
                },
                "message": {
                    "description": "Translated error message"
                }
            }
        },
...
    }
}